kaam chaalu haa

+++ domain/src/main/java/com/example/domain/di/usecase/DeleteProductUseCase.kt
@@ -4,7 +4,7 @@ import com.example.domain.di.model.CartItemModel
 import com.example.domain.di.repo.CartRepository
 
 class DeleteProductUseCase(private val cartRepository: CartRepository) {
-    suspend fun execute(cartItemId: Int, userId: Int) = cartRepository.deleteItem(cartItemId, userId)
+    suspend fun execute(cartItemId: Int, userId: Long) = cartRepository.deleteItem(cartItemId, userId)
 
 }
